+++ BackEndDoctApp/src/main/java/com/example/BackEndDoctApp/entity/Patient.java
@@ -16,7 +16,7 @@ public class Patient {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Long cin;
+    private int  cin;
     private String nom;
     private String prenom;
     private String mail;
